Task module changed version

+++ Create-Task-Application/src/main/java/com/batch5/Create_Task_Application/taskModule/service/TaskService.java
@@ -21,7 +21,7 @@ public interface TaskService {
 
     List<TaskResponseDTO> getTaskByProjectId(Integer projectId);
 
-    List<TaskResponseDTO> getTaskByUserId(Integer userId);
+    List<TaskResponseDTO> getTaskByUserId(Long userId);
 
     List<TaskResponseDTO> getTaskByStatus(String status);
